chore/refactoring back button
