+++ src/components/BackButton/styles.ts
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const BackBtn = styled.button`
+    font-family: Saira;
+    color: #617480;
+
+    border: none;
+    margin-bottom: 2.2rem;
+    background-color: transparent;
+    cursor: pointer;
+    
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+
+    svg {
+        font-size: 1.8rem;
+    }
+
+    p {
+        font-size: 1.4rem;
+    }
+`
